refactor: several refactors (read commit description)
- Move implementation to `src` folder.
- Move dependency index to `pyproject.py`, delete `requirements.txt` then. Also
  pin-point working dependency version in the index.
- Add `mswaa` executable in the `pyproject.py` so it gets installed when the
  package is installed.
- Update `README.md` with newer and simpler installation instructions.
- Get rid of shell scripts like `setup.sh` and `main.sh` since they're useless
now.
