+++ .gitignore
@@ -2,3 +2,5 @@
 __pycache__
 .vscode/
 venv/
+.venv/
+.ruff_cache/
