--- setup.sh
@@ -1,7 +1,0 @@
-#! /usr/bin/env bash
-
-cwd=$(dirname $(readlink -f ${BASH_SOURCE[0]}))
-
-python3 -m venv venv
-source venv/bin/activate
-python3 -m pip install -r requirements.txt
