+++ src/mswaa/common.py
@@ -38,8 +38,6 @@ def install_default_config():
     if not config_dir.exists():
         os.makedirs(config_dir)
     with open(config_fn, "w") as f:
-        f.write(
-            f"# automatically generated by {app_name} install_default_config()\n"
-        )
+        f.write(f"# automatically generated by {app_name} install_default_config()\n")
         for key, value in good_default_options().items():
             f.write(f"{key} = {value!r}\n")
