+++ src/mswaa/vec2.py
